+++ src/eventHandlers/action/removed.ts
