+++ src/plugin.ts
@@ -10,13 +10,13 @@ import vatsimManager from "@managers/vatsim";
 import { AtisLetter } from "@actions/atisLetter";
 import { PushToTalk } from "@actions/pushToTalk";
 import { TrackAudioStatus } from "@actions/trackAudioStatus";
-import { handleAtisLetterAdded } from "@eventHandlers/actionManager/atisLetterAdded";
-import { handleAtisLetterUpdated } from "@eventHandlers/actionManager/atisLetterUpdated";
-import { handleHotlineSettingsUpdated } from "@eventHandlers/actionManager/hotlineSettingsUpdated";
-import { handleRemoved } from "@eventHandlers/actionManager/removed";
-import { handleStationStatusAdded } from "@eventHandlers/actionManager/stationStatusAdded";
-import { handleStationStatusSettingsUpdated } from "@eventHandlers/actionManager/stationStatusSettingsUpdated";
-import { handleTrackAudioStatusAdded } from "@eventHandlers/actionManager/trackAudioStatusAdded";
+import { handleAtisLetterAdded } from "@eventHandlers/action/atisLetterAdded";
+import { handleAtisLetterUpdated } from "@eventHandlers/action/atisLetterUpdated";
+import { handleHotlineSettingsUpdated } from "@eventHandlers/action/hotlineSettingsUpdated";
+import { handleRemoved } from "@eventHandlers/action/removed";
+import { handleStationStatusAdded } from "@eventHandlers/action/stationStatusAdded";
+import { handleStationStatusSettingsUpdated } from "@eventHandlers/action/stationStatusSettingsUpdated";
+import { handleTrackAudioStatusAdded } from "@eventHandlers/action/trackAudioStatusAdded";
 import { handleImageChanged } from "@eventHandlers/svg/imageChanged";
 import { handleConnected } from "@eventHandlers/trackAudio/connected";
 import { handleDisconnected } from "@eventHandlers/trackAudio/disconnected";
